// packages/table/table.type.ts
interface fn {
  (...args: any[]): any;
}

interface AnyObject {
  [key: string]: any;
}
// ...
interface TableColumnCtx {
  id?: string;
  realWidth: number;
  type: string;
  label: string;
  className: string;
  labelClassName: string;
  property: string;
  prop: string;
  width: number;
  minWidth: number;
  renderHeader: fn;
  sortable: boolean | string;
  sortMethod: fn;
  sortBy: string | fn | unknown[];
  resizable: {
    type: boolean;
    default: true;
  };
  columnKey: string;
  align: string;
  headerAlign: string;
  showTooltipWhenOverflow: boolean;
  showOverflowTooltip: boolean;
  fixed: boolean | string;
  formatter: fn;
  selectable: fn;
  reserveSelection: boolean;
  filterMethod: fn;
  filteredValue: unknown[];
  filters: unknown[];
  filterPlacement: string;
  filterMultiple: {
    type: boolean;
    default: true;
  };
  index: number | fn;
  sortOrders: unknown[];
  renderCell: (data: AnyObject) => void;
  colSpan: number;
  rowSpan: number;
  children: TableColumnCtx[];
  level: number;
  filterable: boolean | fn;
  order: string;
  isColumnGroup: boolean;
  columns: TableColumnCtx[];
  getColumnIndex: () => number;
  no: number;
}